[bfops/move-lockfile]: move lockfile
